Added UseEnumSetOf recipe. Added `java-util-apis` declarative recipe. Added `java-util-apis` recipe to java 8 to 11. Added `util` package for `java.util` recipes. Moved `util` recipes from `lang` into `util` package. fixes #88
