+++ src/main/java/org/openrewrite/java/migrate/util/MigrateCollectionsSingletonSet.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.openrewrite.java.migrate.lang;
+package org.openrewrite.java.migrate.util;
 
 import org.openrewrite.ExecutionContext;
 import org.openrewrite.Recipe;
@@ -27,12 +27,12 @@ import org.openrewrite.java.tree.J;
 
 import java.time.Duration;
 
-public class MigrateCollectionsSingletonList extends Recipe {
-    private static final MethodMatcher SINGLETON_LIST = new MethodMatcher("java.util.Collections singletonList(..)", true);
+public class MigrateCollectionsSingletonSet extends Recipe {
+    private static final MethodMatcher SINGLETON_SET = new MethodMatcher("java.util.Collections singleton(..)", true);
 
     @Override
     public String getDisplayName() {
-        return "Use `List.of(..)` in Java 9 or higher";
+        return "Use `Set.of(..)` in Java 9 or higher";
     }
 
     @Override
@@ -42,7 +42,7 @@ public class MigrateCollectionsSingletonList extends Recipe {
 
     @Override
     public String getDescription() {
-        return "Replaces `Collections.singletonList(<args>)))` with `List.Of(<args>)`.";
+        return "Replaces `Collections.singleton(<args>)))` with `Set.Of(<args>)`.";
     }
 
     @Override
@@ -51,7 +51,7 @@ public class MigrateCollectionsSingletonList extends Recipe {
             @Override
             public J.CompilationUnit visitCompilationUnit(J.CompilationUnit cu, ExecutionContext executionContext) {
                 doAfterVisit(new UsesJavaVersion<>(9));
-                doAfterVisit(new UsesMethod<>(SINGLETON_LIST));
+                doAfterVisit(new UsesMethod<>(SINGLETON_SET));
                 return cu;
             }
         };
@@ -63,13 +63,13 @@ public class MigrateCollectionsSingletonList extends Recipe {
             @Override
             public J visitMethodInvocation(J.MethodInvocation method, ExecutionContext executionContext) {
                 J.MethodInvocation m = (J.MethodInvocation) super.visitMethodInvocation(method, executionContext);
-                if (SINGLETON_LIST.matches(method)) {
+                if (SINGLETON_SET.matches(method)) {
                     maybeRemoveImport("java.util.Collections");
 
                     return autoFormat(m.withTemplate(
                             JavaTemplate
-                                    .builder(this::getCursor, "List.of(#{any()})")
-                                    .imports("java.util.List")
+                                    .builder(this::getCursor, "Set.of(#{any()})")
+                                    .imports("java.util.Set")
                                     .build(),
                             m.getCoordinates().replace(),
                             m.getArguments().get(0)
