+++ src/main/java/org/openrewrite/java/migrate/util/MigrateCollectionsUnmodifiableList.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.openrewrite.java.migrate.lang;
+package org.openrewrite.java.migrate.util;
 
 import org.openrewrite.ExecutionContext;
 import org.openrewrite.Recipe;
@@ -23,16 +23,20 @@ import org.openrewrite.java.JavaVisitor;
 import org.openrewrite.java.MethodMatcher;
 import org.openrewrite.java.search.UsesJavaVersion;
 import org.openrewrite.java.search.UsesMethod;
+import org.openrewrite.java.tree.Expression;
 import org.openrewrite.java.tree.J;
 
 import java.time.Duration;
+import java.util.List;
+import java.util.StringJoiner;
 
-public class MigrateCollectionsSingletonSet extends Recipe {
-    private static final MethodMatcher SINGLETON_SET = new MethodMatcher("java.util.Collections singleton(..)", true);
+public class MigrateCollectionsUnmodifiableList extends Recipe {
+    private static final MethodMatcher UNMODIFIABLE_LIST = new MethodMatcher("java.util.Collections unmodifiableList(java.util.List)", true);
+    private static final MethodMatcher ARRAYS_AS_LIST = new MethodMatcher("java.util.Arrays asList(..)", true);
 
     @Override
     public String getDisplayName() {
-        return "Use `Set.of(..)` in Java 9 or higher";
+        return "Use `List.of(..)` in Java 9 or higher";
     }
 
     @Override
@@ -42,7 +46,7 @@ public class MigrateCollectionsSingletonSet extends Recipe {
 
     @Override
     public String getDescription() {
-        return "Replaces `Collections.singleton(<args>)))` with `Set.Of(<args>)`.";
+        return "Replaces `unmodifiableList(java.util.Arrays asList(<args>))` with `List.Of(<args>)`.";
     }
 
     @Override
@@ -51,7 +55,7 @@ public class MigrateCollectionsSingletonSet extends Recipe {
             @Override
             public J.CompilationUnit visitCompilationUnit(J.CompilationUnit cu, ExecutionContext executionContext) {
                 doAfterVisit(new UsesJavaVersion<>(9));
-                doAfterVisit(new UsesMethod<>(SINGLETON_SET));
+                doAfterVisit(new UsesMethod<>(UNMODIFIABLE_LIST));
                 return cu;
             }
         };
@@ -63,19 +67,28 @@ public class MigrateCollectionsSingletonSet extends Recipe {
             @Override
             public J visitMethodInvocation(J.MethodInvocation method, ExecutionContext executionContext) {
                 J.MethodInvocation m = (J.MethodInvocation) super.visitMethodInvocation(method, executionContext);
-                if (SINGLETON_SET.matches(method)) {
-                    maybeRemoveImport("java.util.Collections");
+                if (UNMODIFIABLE_LIST.matches(method)) {
+                    if (m.getArguments().get(0) instanceof J.MethodInvocation) {
+                        if (ARRAYS_AS_LIST.matches((J.MethodInvocation) m.getArguments().get(0))) {
+                            J.MethodInvocation arraysInvocation = (J.MethodInvocation) m.getArguments().get(0);
+                            maybeRemoveImport("java.util.Collections");
+                            maybeRemoveImport("java.util.Arrays");
 
-                    return autoFormat(m.withTemplate(
-                            JavaTemplate
-                                    .builder(this::getCursor, "Set.of(#{any()})")
-                                    .imports("java.util.Set")
-                                    .build(),
-                            m.getCoordinates().replace(),
-                            m.getArguments().get(0)
-                    ), executionContext);
+                            StringJoiner setOf = new StringJoiner(", ", "List.of(", ")");
+                            List<Expression> args = arraysInvocation.getArguments();
+                            args.forEach(o -> setOf.add("#{any()}"));
+
+                            return autoFormat(m.withTemplate(
+                                    JavaTemplate
+                                            .builder(this::getCursor, setOf.toString())
+                                            .imports("java.util.List")
+                                            .build(),
+                                    m.getCoordinates().replace(),
+                                    args.toArray()
+                            ), executionContext);
+                        }
+                    }
                 }
-
                 return m;
             }
         };
