+++ src/main/resources/META-INF/rewrite/java-util-apis.yml
@@ -1,0 +1,26 @@
+#
+# Copyright 2022 the original author or authors.
+# <p>
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+# <p>
+# https://www.apache.org/licenses/LICENSE-2.0
+# <p>
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+---
+type: specs.openrewrite.org/v1beta/recipe
+name: org.openrewrite.java.migrate.util.JavaUtilAPIs
+displayName: Migrate `java.util` APIs
+description: Certain java util APIs have been introduced and are favored over previous APIs.
+recipeList:
+  - org.openrewrite.java.migrate.util.MigrateCollectionsSingletonList
+  - org.openrewrite.java.migrate.util.MigrateCollectionsSingletonMap
+  - org.openrewrite.java.migrate.util.MigrateCollectionsSingletonSet
+  - org.openrewrite.java.migrate.util.MigrateCollectionsUnmodifiableList
+  - org.openrewrite.java.migrate.util.MigrateCollectionsUnmodifiableSet
