+++ src/main/resources/META-INF/rewrite/java8-to-java11.yml
@@ -46,3 +46,4 @@ recipeList:
   - org.openrewrite.java.migrate.wro4j.UpgradeWro4jMavenPluginVersion
   - org.openrewrite.java.migrate.jacoco.UpgradeJaCoCoMavenPluginVersion
   - org.openrewrite.java.migrate.JavaVersion11
+  - org.openrewrite.java.migrate.util.JavaUtilAPIs
