+++ src/test/kotlin/org/openrewrite/java/migrate/util/MigrateCollectionsSingletonMapTest.kt
@@ -13,53 +13,53 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.openrewrite.java.migrate.lang
+package org.openrewrite.java.migrate.util
 
 import org.junit.jupiter.api.Test
 import org.openrewrite.Issue
 import org.openrewrite.Recipe
 import org.openrewrite.java.JavaRecipeTest
 
-class MigrateCollectionsSingletonSetTest : JavaRecipeTest {
+class MigrateCollectionsSingletonMapTest : JavaRecipeTest {
     override val recipe: Recipe
-        get() = MigrateCollectionsSingletonSet()
+        get() = MigrateCollectionsSingletonMap()
 
     @Issue("https://github.com/openrewrite/rewrite-migrate-java/issues/72")
     @Test
-    fun singleTonSet() = assertChanged(
+    fun singletonMap() = assertChanged(
         before = """
             import java.util.*;
             
             class Test {
-                Set<String> set = Collections.singleton("Hello");
+                Map<String,String> set = Collections.singletonMap("hello", "world");
             }
         """,
         after = """
-            import java.util.Set;
+            import java.util.Map;
             
             class Test {
-                Set<String> set = Set.of("Hello");
+                Map<String,String> set = Map.of("hello", "world");
             }
         """
     )
 
     @Issue("https://github.com/openrewrite/rewrite-migrate-java/issues/72")
     @Test
-    fun singletonSetCustomType() = assertChanged(
+    fun singletonMapCustomType() = assertChanged(
         before = """
             import java.util.*;
             import java.time.LocalDate;
             
             class Test {
-                Set<LocalDate> set = Collections.singleton(LocalDate.now());
+                Map<String,LocalDate> map = Collections.singletonMap("date", LocalDate.now());
             }
         """,
         after = """
-            import java.util.Set;
+            import java.util.Map;
             import java.time.LocalDate;
             
             class Test {
-                Set<LocalDate> set = Set.of(LocalDate.now());
+                Map<String,LocalDate> map = Map.of("date", LocalDate.now());
             }
         """
     )
